fix: change auction bet type

+++ db/migrations/sqls/20190822154502-create-table-auction-up.sql
@@ -2,7 +2,7 @@ CREATE TABLE "auction" (
   "auction_id" SERIAL      NOT NULL,
   "user_id"    INTEGER     NOT NULL  REFERENCES "users"("user_id"),
   "status"     VARCHAR(25),
-  "bet"        INTEGER     NOT NULL,
+  "bet"        FLOAT       NOT NULL,
   "time"       TIMESTAMP WITHOUT TIME ZONE DEFAULT now(),
 
   PRIMARY KEY("auction_id")
